+++ example/lib/main.dart
@@ -15,11 +15,24 @@ class MyApp extends StatelessWidget {
     return MaterialApp(
       title: 'Flutter Demo',
       theme: ThemeData(
-        primarySwatch: Colors.blue,
+        // This is the theme of your application.
+        //
+        // TRY THIS: Try running your application with "flutter run". You'll see
+        // the application has a purple toolbar. Then, without quitting the app,
+        // try changing the seedColor in the colorScheme below to Colors.green
+        // and then invoke "hot reload" (save your changes or press the "hot
+        // reload" button in a Flutter-supported IDE, or press "r" if you used
+        // the command line to start the app).
+        //
+        // Notice that the counter didn't reset back to zero; the application
+        // state is not lost during the reload. To reset the state, use hot
+        // restart instead.
+        //
+        // This works for code too, not just values: Most code changes can be
+        // tested with just a hot reload.
+        colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
       ),
-      home: const MyHomePage(
-        title: 'Gif View Example',
-      ),
+      home: const MyHomePage(title: 'Gif View Example'),
     );
   }
 }
@@ -57,9 +70,9 @@ class _PreCachePageState extends State<PreCachePage> {
                       Navigator.push(
                         context,
                         MaterialPageRoute(
-                          builder: (context) => const MyHomePage(
-                            title: 'Gif View Example',
-                          ),
+                          builder:
+                              (context) =>
+                                  const MyHomePage(title: 'Gif View Example'),
                         ),
                       );
                     },
@@ -96,9 +109,9 @@ class _PreCachePageState extends State<PreCachePage> {
 
   Future<void> _preCacheGif() async {
     for (final gif in gifs) {
-      ImageProvider provider = (gif.startsWith('http')
-          ? NetworkImage(gif)
-          : AssetImage(gif)) as ImageProvider;
+      ImageProvider provider =
+          (gif.startsWith('http') ? NetworkImage(gif) : AssetImage(gif))
+              as ImageProvider;
       gifLoading.value = gif;
       await GifView.preFetchImage(provider);
     }
@@ -115,13 +128,10 @@ class MyHomePage extends StatefulWidget {
 }
 
 class _MyHomePageState extends State<MyHomePage> {
-  
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(
-        title: Text(widget.title),
-      ),
+      appBar: AppBar(title: Text(widget.title)),
       body: ListView(
         children: [
           Padding(
@@ -132,11 +142,7 @@ class _MyHomePageState extends State<MyHomePage> {
             ),
           ),
           const Divider(),
-          GifView.asset(
-            gifs[0],
-            height: 200,
-            frameRate: 30,
-          ),
+          GifView.asset(gifs[0], height: 200, frameRate: 30),
           Padding(
             padding: const EdgeInsets.all(8.0),
             child: Text(
@@ -150,16 +156,13 @@ class _MyHomePageState extends State<MyHomePage> {
             shrinkWrap: true,
             physics: const NeverScrollableScrollPhysics(),
             children: [
-              GifView.network(
-                gifs[1],
-                height: 200,
-              ),
+              GifView.network(gifs[1], height: 200),
               GifView.network(
                 gifs[2],
                 height: 200,
-                progressBuilder: (context) => const Center(
-                  child: CircularProgressIndicator(),
-                ),
+                progressBuilder:
+                    (context) =>
+                        const Center(child: CircularProgressIndicator()),
                 errorBuilder: (context, error, tryAgain) {
                   return InkWell(
                     onTap: tryAgain,
@@ -173,10 +176,7 @@ class _MyHomePageState extends State<MyHomePage> {
                   print('onStart');
                 },
               ),
-              GifView.network(
-                gifs[3],
-                height: 200,
-              ),
+              GifView.network(gifs[3], height: 200),
             ],
           ),
         ],